+++ src/Configurations.ts
@@ -16,8 +16,6 @@ import { platformUtil } from './util/Platform';
 import { cloneRecursively } from './util/ResolveRule';
 import { DebugConfigData } from './DebugConfigType';
 
-type SentryValue = 'question' | 'enable' | 'enabled' | 'disable' | 'disable_1' | 'disable_2' | 'disable_3';
-
 const ConfigSectionBase = 'testMate.cpp';
 
 const enum Section {
@@ -48,7 +46,6 @@ export type Config =
   | 'debug.noThrow'
   | 'log.logpanel'
   | 'log.logfile'
-  | 'log.logSentry'
   | 'log.userId'
   | 'gtest.treatGmockWarningAs'
   | 'gtest.gmockVerbose';
@@ -314,60 +311,6 @@ export class Configurations {
   //   return decrypted.toString('utf8');
   // }
 
-  isSentryEnabled(): boolean {
-    const val = this._get('log.logSentry');
-    return val === 'enable' || val === 'enabled';
-  }
-
-  askSentryConsent(): void {
-    const envAskSentry = process.env['TESTMATE_CPP_ASKSENTRYCONSENT'];
-    if (envAskSentry === 'disabled_3') {
-      return;
-      //const decrypted = Configurations.decrypt(process.env['TESTMATE_CPP_LOGSENTRY']);
-      //if (decrypted === 'disable_3') return;
-    }
-
-    const logSentryConfig: Config = 'log.logSentry';
-
-    const logSentry = this._getD<SentryValue>(logSentryConfig, 'question');
-
-    if (logSentry === 'question' || logSentry === 'disable' || logSentry === 'disable_1' || logSentry === 'disable_2') {
-      const options = [
-        'Sure! I love this extension and happy to help.',
-        'Yes, but exclude current workspace',
-        'Over my dead body (No)',
-      ];
-      vscode.window
-        .showInformationMessage(
-          'Hey there! C++ TestMate has [sentry.io](https://github.com/matepek/vscode-catch2-test-adapter/blob/master/documents/configuration/log.logSentry.md) integration to ' +
-            'improve the stability and the development. 🤩 For this I want to send logs and errors ' +
-            'but I would NEVER do it without your consent. ' +
-            'Please be understandable and allow it. 🙏',
-          ...options,
-        )
-        .then((value: string | undefined) => {
-          this._log.info('Sentry consent', value);
-
-          if (value === options[0]) {
-            this._cfg
-              .update(logSentryConfig, 'enable', vscode.ConfigurationTarget.Global)
-              .then(undefined, e => this._log.exceptionS(e));
-          } else if (value === options[1]) {
-            this._cfg
-              .update(logSentryConfig, 'enable', vscode.ConfigurationTarget.Global)
-              .then(undefined, e => this._log.exceptionS(e));
-            this._cfg
-              .update(logSentryConfig, 'disable_3', vscode.ConfigurationTarget.WorkspaceFolder)
-              .then(undefined, e => this._log.exceptionS(e));
-          } else if (value === options[2]) {
-            this._cfg
-              .update(logSentryConfig, 'disable_3', vscode.ConfigurationTarget.Global)
-              .then(undefined, e => this._log.exceptionS(e));
-          }
-        });
-    }
-  }
-
   getLoadAtStartup(): boolean {
     return this._getD<boolean>('discovery.loadOnStartup', false);
   }
@@ -482,7 +425,7 @@ export class Configurations {
         undefined,
         undefined,
         undefined,
-        undefined,
+        false,
         undefined,
         undefined,
         {},
@@ -549,6 +492,7 @@ export class Configurations {
 
         const description: string | undefined = typeof obj.description === 'string' ? obj.description : undefined;
 
+        // eslint-disable-next-line no-useless-assignment
         let pattern = '';
         {
           if (typeof obj.pattern == 'string') pattern = obj.pattern;
@@ -603,7 +547,7 @@ export class Configurations {
 
         const executableSuffixToInclude: string[] | undefined = obj.executableSuffixToInclude;
 
-        const waitForBuildProcess: boolean | string | undefined = obj.waitForBuildProcess;
+        const waitForBuildProcess: boolean | string = obj.waitForBuildProcess ?? false;
 
         const debugConfigData: DebugConfigData | undefined = obj['debug.configTemplate']
           ? this._getDebugConfigData(obj['debug.configTemplate'])
@@ -612,12 +556,7 @@ export class Configurations {
         const defaultTestGrouping = obj.testGrouping;
 
         const spawnerConfig: ExecutionWrapperConfig | undefined =
-          typeof obj.executionWrapper === 'object' &&
-          typeof obj.executionWrapper.path === 'string' &&
-          (obj.executionWrapper.args === undefined ||
-            (Array.isArray(obj.executionWrapper.args) && obj.executionWrapper.args.every(x => typeof x === 'string')))
-            ? obj.executionWrapper
-            : undefined;
+          typeof obj.executionWrapper === 'object' ? obj.executionWrapper : undefined;
 
         const sourceFileMap: Record<string, string> =
           typeof obj.sourceFileMap === 'object' &&