+++ src/util/BuildProcessChecker.ts
@@ -1,73 +1,124 @@
 import { Logger } from '../Logger';
-import find from 'find-process';
+import findProcess from 'find-process';
+import { promisify } from 'node:util';
+import { CancellationToken } from '../Util';
 
 ///
 
-// not so nice, init in rootsuite in the future
-export class BuildProcessChecker {
-  constructor(private readonly _log: Logger) {}
+export interface BuildProcessChecker {
+  dispose(): void;
+  resolveAtFinish(pattern: string | boolean | undefined, token: CancellationToken): Promise<void>;
+}
 
-  private readonly _checkIntervalMillis = 2000;
-  // https://en.wikipedia.org/wiki/List_of_compilers#C++_compilers
-  private readonly _defaultPattern =
-    /(^|[/\\])(bazel|cmake|make|ninja|cl|c\+\+|ld|clang|clang\+\+|gcc|g\+\+|link|icc|armcc|armclang)(-[^/\\]+)?(\.exe)?$/;
-  private _lastChecked = 0;
-  private _finishedP = Promise.resolve();
-  private _finishedResolver = (): void => {};
-  private _timerId: NodeJS.Timeout | undefined = undefined; // number if have running build process
+///
+
+const _checkIntervalMillis = 2000;
+// https://en.wikipedia.org/wiki/List_of_compilers#C++_compilers
+const _defaultPattern =
+  /(^|[/\\])(bazel|cmake|make|ninja|cl|c\+\+|ld|clang|clang\+\+|gcc|g\+\+|link|icc|armcc|armclang)(-[^/\\]+)?(\.exe)?$/;
+
+///
+
+export abstract class BuildProcessCheckerBase {
+  constructor(protected readonly _log: Logger) {}
+
+  private _isShutdown = false;
 
   dispose(): void {
-    if (this._timerId) clearInterval(this._timerId);
-    this._finishedResolver();
+    this._isShutdown = true;
   }
 
-  resolveAtFinish(pattern: string | boolean | undefined): Promise<void> {
+  private _patternToUseCache: { patternToUse: RegExp; pattern: string } | undefined = undefined;
+  private readonly _runningChecks = new Map<string | boolean, Promise<void>>();
+
+  resolveAtFinish(pattern: string | boolean, token: CancellationToken): Promise<void> {
     if (pattern === false) return Promise.resolve();
 
-    if (this._timerId !== undefined) {
-      return this._finishedP;
+    let p = this._runningChecks.get(pattern);
+
+    if (p === undefined) {
+      let patternToUse;
+      if (typeof pattern === 'string') {
+        if (this._patternToUseCache?.pattern === pattern) {
+          patternToUse = this._patternToUseCache.patternToUse;
+        } else {
+          patternToUse = new RegExp(pattern);
+          this._patternToUseCache = { pattern, patternToUse };
+        }
+      } else {
+        patternToUse = _defaultPattern;
+      }
+
+      p = Promise.resolve()
+        .then(async () => {
+          while (!this._isShutdown && !token.isCancellationRequested) {
+            try {
+              const processes = await this._find(patternToUse);
+              if (processes.length > 0) {
+                this._log.info(
+                  'Found running build related processes: ' +
+                    processes.map(x => JSON.stringify(x, undefined, 0)).join(', '),
+                );
+              } else {
+                this._log.debug('Not found running build related process');
+                return;
+              }
+            } catch (reason) {
+              this._log.exceptionS('Finding process', reason);
+              return;
+            }
+            await promisify(setTimeout)(_checkIntervalMillis);
+          }
+        })
+        .finally(() => {
+          // for a short period of time we just assume that no new build process was spawned
+          promisify(setTimeout)(300).then(() => this._runningChecks.delete(pattern));
+        });
+
+      this._runningChecks.set(pattern, p);
+
+      this._log.info('Checking running build related processes', patternToUse);
     }
 
-    const elapsed = Date.now() - this._lastChecked;
-
-    if (elapsed < 300) {
-      return Promise.resolve();
-    }
-
-    this._finishedP = new Promise(r => {
-      this._finishedResolver = r;
-    });
-
-    const patternToUse = typeof pattern == 'string' ? RegExp(pattern) : this._defaultPattern;
-    this._log.info('Checking running build related processes', patternToUse);
-    this._timerId = global.setInterval(this._refresh.bind(this, patternToUse), this._checkIntervalMillis);
-    this._refresh(patternToUse);
-
-    return this._finishedP;
+    return p;
   }
 
-  private async _refresh(pattern: RegExp): Promise<void> {
-    try {
-      // wrong type definition for find habdles RegExp: https://github.com/yibn2008/find-process/compare/1.4.11...2.0.0#diff-81b33228621820bded04ffbd7d49375fc742662fde6b7111ddb10457ceef7ae9R11
-      const processes = await find('name', pattern as unknown as string);
+  protected abstract _find(pattern: RegExp): Promise<string[]>;
+}
 
-      this._lastChecked = Date.now();
+///
 
-      if (processes.length > 0) {
-        this._log.info(
-          'Found running build related processes: ' + processes.map(x => JSON.stringify(x, undefined, 0)).join(', '),
-        );
-      } else {
-        this._log.info('Not found running build related process');
-        this._finishedResolver();
-        clearInterval(this._timerId!);
-        this._timerId = undefined;
-      }
-    } catch (reason) {
-      this._log.exceptionS(reason);
-      clearInterval(this._timerId!);
-      this._timerId = undefined;
-      this._finishedResolver();
-    }
+export class FindProcessChecker extends BuildProcessCheckerBase {
+  protected override async _find(pattern: RegExp): Promise<string[]> {
+    const processes = await findProcess('name', pattern);
+    return processes.map(p => p.name);
   }
 }
+
+///
+
+// import psList, { ProcessDescriptor } from 'ps-list';
+// export class PSListProcessChecker extends BuildProcessCheckerBase {
+//   private runningFind: Promise<ProcessDescriptor[]> | undefined = undefined;
+
+//   protected override async _find(pattern: RegExp): Promise<string[]> {
+//     throw Error('https://github.com/matepek/vscode-catch2-test-adapter/issues/524');
+//     if (this.runningFind) {
+//       const ps = await this.runningFind;
+//       return ps.filter(p => p.name.match(pattern)).map(p => p.name);
+//     }
+//     this.runningFind = psList({ all: false });
+//     const ps = await this.runningFind;
+//     this.runningFind = undefined;
+//     return ps.filter(p => p.name.match(pattern)).map(p => p.name);
+//   }
+// }
+
+export const buildProcessCheckerFactory = {
+  // https://www.npmjs.com/package/ps-list : "Works on macOS, Linux, and Windows. Windows ARM64 is not supported yet."
+  create: (log: Logger) => {
+    return /*process.platform === 'win32' && process.arch == 'x64' ? new PSListProcessChecker(log) : */ new FindProcessChecker(
+      log,
+    );
+  },
+};